Us rank.1

+++ src/Controller/DashboardController.php
@@ -32,8 +32,9 @@ class DashboardController extends AbstractController
     {
         $errors = [];
 
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            // clean $_POST data
+        if (
+            !empty($_POST['title']) && !empty($_POST['category']) && !empty($_POST['artist'])
+        ) {
             $item = array_map('trim', $_POST);
 
             $fileName =  basename($_FILES['path']['name']);
@@ -46,8 +47,6 @@ class DashboardController extends AbstractController
             if ((!in_array($extension, $authorizedExtensions))) {
                 $errors[] = 'Veuillez sélectionner un morceau en mp3!';
             }
-
-            // $uploadFile = __DIR__ . '/../../public/uploads' . $fileName;
 
             if (file_exists($_FILES['path']['tmp_name']) && filesize($_FILES['path']['tmp_name']) > $maxFileSize) {
                 $errors[] = "Votre fichier doit faire moins de 5Mo !";
@@ -63,10 +62,11 @@ class DashboardController extends AbstractController
                 $item['path'] = $fileName;
                 $track = $trackManager->insert($item);
                 $answerManager->insert($item, $track);
-                header('Location: /dashboard');
+                header('Location: /dashboard/show?id=' . $track);
                 return null;
             }
         }
+        $errors[] = "Veuillez renseigner tous les champs";
         return $this->twig->render('Dashboard/dashboard.html.twig', ['errors' => $errors]);
     }
 
@@ -99,9 +99,10 @@ class DashboardController extends AbstractController
         $id = trim($id);
         $trackManager = new TrackManager();
         $delete = $trackManager->selectOneById(intVal($id));
-        unlink(__DIR__ . '/../../public/uploads/tracks/' . $delete['path']);
+        if (file_exists(__DIR__ . '/../../public/uploads/tracks/' . $delete['path'])) {
+            unlink(__DIR__ . '/../../public/uploads/tracks/' . $delete['path']);
+        }
         $trackManager->delete((int)$id);
-
         header('Location:/dashboard');
     }
 }
